Carefully check for site packages without import.
Tango app building must not hide ImportErrors which are present in the
site package.  Supporting flexible command-line parsing however requires
an import check to allow both implicit and explicit site package names.

This supports both `tango subcommand default` and
`tango subcommand tango.site.default` which allows for conventional and
unconventional tango site packaging, in terms of Tango conventions.
This is Python, and we intend to support it.