Touch up snapshot usage message.
At this point, snapshot should be generally useful to create routes,
both application setup and template context, which are created on one
host and loaded on another, without the original code around.  Of
course, tango has import name requirements, and in general this approach
is a bad idea because bugs would be difficult to identify and fix.